requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
motor==3.1.1
python-dotenv==1.0.0
pydantic==2.0.2
bcrypt==4.0.1
email-validator==2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
httpx==0.25.0
